{"schema":"decodex/commit/1","summary":"Split Decodex app snapshot models","authority":"XY-1115"}
