+++ apps/decodex-app/Sources/DecodexApp/OperatorSnapshotModels.swift
@@ -1,7 +1,5 @@
 import Foundation
 
-private let operatorCurrentLaneIdleTimeoutSeconds = 300
-
 struct OperatorSnapshotResponse: Decodable, Sendable {
 	let warnings: [String]
 	let projects: [OperatorProjectStatus]
@@ -70,7 +68,7 @@ struct OperatorSnapshotResponse: Decodable, Sendable {
 	var warningSummary: String? {
 		let labels = warnings
 			.filter { $0 != "automation_disabled" }
-			.map(rawDisplayToken)
+			.map(operatorRawDisplayToken)
 			.filter { $0.isEmpty == false }
 		guard let first = labels.first else {
 			return nil
@@ -261,869 +259,4 @@ struct OperatorPostReviewLaneStatus: Decodable, Sendable {
 		classification = try container.decodeIfPresent(String.self, forKey: .classification)
 		shadowedByCurrentLane = try container.decode(Bool.self, forKey: .shadowedByCurrentLane)
 	}
-}
-
-struct OperatorLifecycleMetricBucket: Decodable, Sendable {
-	let name: String
-	let wallSeconds: Int
-	let eventCount: Int
-	let toolCallCount: Int
-	let inputTokens: Int
-	let outputTokens: Int
-	let outputBytes: Int
-
-	enum CodingKeys: String, CodingKey {
-		case name
-		case wallSeconds = "wall_seconds"
-		case eventCount = "event_count"
-		case toolCallCount = "tool_call_count"
-		case inputTokens = "input_tokens"
-		case outputTokens = "output_tokens"
-		case outputBytes = "output_bytes"
-	}
-}
-
-struct OperatorLifecycleMetricPhase: Decodable, Sendable {
-	let phase: String?
-	let label: String?
-	let attemptCount: Int
-	let runCount: Int
-	let capturedAttemptCount: Int
-	let missingAttemptCount: Int
-	let protocolEventCount: Int
-	let childEventCount: Int
-	let wallSeconds: Int
-	let toolCallCount: Int
-	let inputTokensCurrent: Int?
-	let inputTokensPeak: Int?
-	let inputTokensCumulative: Int
-	let outputTokensCumulative: Int
-	let largestToolOutputBytes: Int?
-	let largestToolOutputTool: String?
-	let buckets: [OperatorLifecycleMetricBucket]
-
-	enum CodingKeys: String, CodingKey {
-		case phase
-		case label
-		case attemptCount = "attempt_count"
-		case runCount = "run_count"
-		case capturedAttemptCount = "captured_attempt_count"
-		case missingAttemptCount = "missing_attempt_count"
-		case protocolEventCount = "protocol_event_count"
-		case childEventCount = "child_event_count"
-		case wallSeconds = "wall_seconds"
-		case toolCallCount = "tool_call_count"
-		case inputTokensCurrent = "input_tokens_current"
-		case inputTokensPeak = "input_tokens_peak"
-		case inputTokensCumulative = "input_tokens_cumulative"
-		case outputTokensCumulative = "output_tokens_cumulative"
-		case largestToolOutputBytes = "largest_tool_output_bytes"
-		case largestToolOutputTool = "largest_tool_output_tool"
-		case buckets
-	}
-}
-
-struct OperatorLifecycleMetrics: Decodable, Sendable {
-	let attemptCount: Int
-	let runCount: Int
-	let capturedAttemptCount: Int
-	let missingAttemptCount: Int
-	let protocolEventCount: Int
-	let childEventCount: Int
-	let wallSeconds: Int
-	let toolCallCount: Int
-	let inputTokensCurrent: Int?
-	let inputTokensPeak: Int?
-	let inputTokensCumulative: Int
-	let outputTokensCumulative: Int
-	let largestToolOutputBytes: Int?
-	let largestToolOutputTool: String?
-	let buckets: [OperatorLifecycleMetricBucket]
-	let phases: [OperatorLifecycleMetricPhase]
-
-	enum CodingKeys: String, CodingKey {
-		case attemptCount = "attempt_count"
-		case runCount = "run_count"
-		case capturedAttemptCount = "captured_attempt_count"
-		case missingAttemptCount = "missing_attempt_count"
-		case protocolEventCount = "protocol_event_count"
-		case childEventCount = "child_event_count"
-		case wallSeconds = "wall_seconds"
-		case toolCallCount = "tool_call_count"
-		case inputTokensCurrent = "input_tokens_current"
-		case inputTokensPeak = "input_tokens_peak"
-		case inputTokensCumulative = "input_tokens_cumulative"
-		case outputTokensCumulative = "output_tokens_cumulative"
-		case largestToolOutputBytes = "largest_tool_output_bytes"
-		case largestToolOutputTool = "largest_tool_output_tool"
-		case buckets
-		case phases
-	}
-}
-
-struct OperatorRunStatus: Decodable, Identifiable, Sendable {
-	let projectID: String?
-	let projectDisplayName: String?
-	let runID: String
-	let issueID: String?
-	let issueIdentifier: String?
-	let title: String?
-	let status: String?
-	let attemptStatus: String?
-	let attemptNumber: Int?
-	let phase: String?
-	let runPhase: String?
-	let waitReason: String?
-	let currentOperation: String?
-	let activeGoalPhase: String?
-	let publicProgressPhase: String?
-	let threadStatus: String?
-	let threadActiveFlags: [String]
-	let idleForSeconds: Int?
-	let protocolIdleForSeconds: Int?
-	let updatedAt: String?
-	let lastProgressAt: String?
-	let nextRetryAt: String?
-	let lastEventType: String?
-	let eventCount: Int?
-	let executionLiveness: String?
-	let hasFreshExecutionSnapshot: Bool?
-	let countsAsRunningSnapshot: Bool?
-	let needsAttentionSnapshot: Bool?
-	let processAlive: Bool?
-	let processLivenessReason: String?
-	let runLease: Bool?
-	let branchName: String?
-	let worktreePath: String?
-	let suspectedStall: Bool
-	let childAgentActivity: OperatorChildAgentActivity?
-	let lifecycleMetrics: OperatorLifecycleMetrics?
-	let continuationRecovery: OperatorContinuationRecoveryStatus?
-	let account: OperatorRunAccountSummary?
-	let accounts: [OperatorRunAccountSummary]
-
-	var id: String {
-		runID
-	}
-
-	var compactTitle: String {
-		if let issueIdentifier = trimmed(issueIdentifier), issueIdentifier.isEmpty == false {
-			return issueIdentifier
-		}
-		if let title = trimmed(title), title.isEmpty == false {
-			return title
-		}
-
-		return "Run"
-	}
-
-	var compactDetail: String {
-		if let currentDetail = trimmed(childAgentActivity?.currentDetail), currentDetail.isEmpty == false {
-			return currentDetail
-		}
-		if let currentBucket = trimmed(childAgentActivity?.currentBucket), currentBucket.isEmpty == false {
-			return rawDisplayToken(currentBucket)
-		}
-		if let waitReason = trimmed(waitReason), waitReason.isEmpty == false {
-			return rawDisplayToken(waitReason)
-		}
-		if let operation = trimmed(currentOperation), operation.isEmpty == false, operation != "idle" {
-			return rawDisplayToken(operation)
-		}
-		if let runPhase = trimmed(runPhase ?? phase), runPhase.isEmpty == false {
-			return rawDisplayToken(runPhase)
-		}
-		if let threadStatus = trimmed(threadStatus), threadStatus.isEmpty == false {
-			return rawDisplayToken(threadStatus)
-		}
-		if let status = trimmed(status), status.isEmpty == false {
-			return rawDisplayToken(status)
-		}
-
-		return "Active"
-	}
-
-	func compactActivitySummary(at now: Date) -> String? {
-		guard let activity = childAgentActivity else {
-			return nil
-		}
-
-		var parts = [String]()
-		if let modelBucket = activity.buckets.first(where: { $0.name.caseInsensitiveCompare("Model") == .orderedSame }) {
-			let modelSeconds = activity.wallSeconds(for: modelBucket, at: now)
-			if modelSeconds > 0, let formatted = formatOperatorActivityDuration(modelSeconds) {
-				parts.append("Model \(formatted)")
-			}
-		} else if activity.wallSeconds(at: now) > 0,
-			let formatted = formatOperatorActivityDuration(activity.wallSeconds(at: now))
-		{
-			parts.append("Activity \(formatted)")
-		}
-
-		if activity.inputTokensCumulative > 0 || activity.outputTokensCumulative > 0 {
-			parts.append(
-				"in \(formatOperatorCompactCount(activity.inputTokensCumulative)) / out \(formatOperatorCompactCount(activity.outputTokensCumulative))"
-			)
-		}
-		if activity.toolCallCount > 0 {
-			parts.append("\(formatOperatorCompactCount(activity.toolCallCount)) tools")
-		}
-		if let largestOutput = activity.largestToolOutputBytes, largestOutput > 0 {
-			parts.append("\(formatOperatorCompactBytes(largestOutput)) output")
-		}
-
-		return parts.isEmpty ? nil : parts.joined(separator: " · ")
-	}
-
-	var hasAttentionTone: Bool {
-		if let needsAttentionSnapshot {
-			return needsAttentionSnapshot
-		}
-
-		return suspectedStall
-			|| attemptStatus == "waiting_for_review"
-			|| status == "manual_attention"
-			|| status == "blocked"
-			|| stoppedProcessNeedsAttention
-	}
-
-	var isWaiting: Bool {
-		waitReason != nil
-			|| attemptStatus?.contains("waiting") == true
-			|| phase?.contains("waiting") == true
-	}
-
-	var inactiveDurationSeconds: Int? {
-		let candidates = [
-			idleForSeconds,
-			protocolIdleForSeconds,
-			childAgentActivity?.currentElapsedSeconds,
-		].compactMap { $0 }
-
-		return candidates.max()
-	}
-
-	func isAssigned(to account: CodexAccount) -> Bool {
-		self.account?.matches(account) == true
-			|| accounts.contains { $0.isSelected && $0.matches(account) }
-	}
-
-	var countsAsRunning: Bool {
-		if let countsAsRunningSnapshot {
-			return countsAsRunningSnapshot
-		}
-
-		return hasRunningStatus
-			&& phase == "executing"
-			&& (processAlive != false || hasFreshExecution)
-			&& hasAttentionTone == false
-			&& hasStaleExecutionWithoutKnownProcess == false
-	}
-
-	var hasFreshExecution: Bool {
-		if let hasFreshExecutionSnapshot {
-			return hasFreshExecutionSnapshot
-		}
-
-		return hasRunningStatus
-			&& (processAlive == true || hasRecentAppServerExecution)
-	}
-
-	private var hasRecentAppServerExecution: Bool {
-		threadStatus == "active"
-			|| threadActiveFlags.isEmpty == false
-			|| ["thread_active", "protocol_observed"].contains(executionLiveness ?? "")
-			|| protocolIdleForSeconds.isSomeAndLessThan(operatorCurrentLaneIdleTimeoutSeconds)
-	}
-
-	private var hasRunningStatus: Bool {
-		guard let status else {
-			return false
-		}
-
-		return ["starting", "running"].contains(status)
-	}
-
-	private var stoppedProcessNeedsAttention: Bool {
-		processAlive == false
-			&& hasRunningStatus
-			&& waitReason == nil
-			&& hasFreshExecution == false
-	}
-
-	private var hasStaleExecutionWithoutKnownProcess: Bool {
-		hasRunningStatus
-			&& phase == "executing"
-			&& waitReason == nil
-			&& processAlive != true
-			&& [idleForSeconds, protocolIdleForSeconds].contains { idleForSeconds in
-				guard let idleForSeconds else {
-					return false
-				}
-
-				return idleForSeconds >= 300
-			}
-	}
-
-	enum CodingKeys: String, CodingKey {
-		case projectID = "project_id"
-		case projectDisplayName = "project_display_name"
-		case runID = "run_id"
-		case issueID = "issue_id"
-		case issueIdentifier = "issue_identifier"
-		case title
-		case status
-		case attemptStatus = "attempt_status"
-		case attemptNumber = "attempt_number"
-		case phase
-		case runPhase = "run_phase"
-		case waitReason = "wait_reason"
-		case currentOperation = "current_operation"
-		case activeGoalPhase = "active_goal_phase"
-		case publicProgressPhase = "public_progress_phase"
-		case threadStatus = "thread_status"
-		case threadActiveFlags = "thread_active_flags"
-		case idleForSeconds = "idle_for_seconds"
-		case protocolIdleForSeconds = "protocol_idle_for_seconds"
-		case updatedAt = "updated_at"
-		case lastProgressAt = "last_progress_at"
-		case nextRetryAt = "next_retry_at"
-		case lastEventType = "last_event_type"
-		case eventCount = "event_count"
-		case executionLiveness = "execution_liveness"
-		case hasFreshExecutionSnapshot = "has_fresh_execution"
-		case countsAsRunningSnapshot = "counts_as_running"
-		case needsAttentionSnapshot = "needs_attention"
-		case processAlive = "process_alive"
-		case processLivenessReason = "process_liveness_reason"
-		case runLease = "run_lease"
-		case branchName = "branch_name"
-		case worktreePath = "worktree_path"
-		case suspectedStall = "suspected_stall"
-		case childAgentActivity = "child_agent_activity"
-		case lifecycleMetrics = "lifecycle_metrics"
-		case continuationRecovery = "continuation_recovery"
-		case account
-		case accounts
-		case codexAccount = "codex_account"
-		case codexAccounts = "codex_accounts"
-	}
-
-	init(from decoder: Decoder) throws {
-		let container = try decoder.container(keyedBy: CodingKeys.self)
-
-		projectID = try container.decodeIfPresent(String.self, forKey: .projectID)
-		projectDisplayName = try container.decodeIfPresent(String.self, forKey: .projectDisplayName)
-		runID = try container.decodeIfPresent(String.self, forKey: .runID) ?? UUID().uuidString
-		issueID = try container.decodeIfPresent(String.self, forKey: .issueID)
-		issueIdentifier = try container.decodeIfPresent(String.self, forKey: .issueIdentifier)
-		title = try container.decodeIfPresent(String.self, forKey: .title)
-		status = try container.decodeIfPresent(String.self, forKey: .status)
-		attemptStatus = try container.decodeIfPresent(String.self, forKey: .attemptStatus)
-		attemptNumber = try container.decodeIfPresent(Int.self, forKey: .attemptNumber)
-		phase = try container.decodeIfPresent(String.self, forKey: .phase)
-		runPhase = try container.decodeIfPresent(String.self, forKey: .runPhase)
-		waitReason = try container.decodeIfPresent(String.self, forKey: .waitReason)
-		currentOperation = try container.decodeIfPresent(String.self, forKey: .currentOperation)
-		activeGoalPhase = try container.decodeIfPresent(String.self, forKey: .activeGoalPhase)
-		publicProgressPhase = try container.decodeIfPresent(String.self, forKey: .publicProgressPhase)
-		threadStatus = try container.decodeIfPresent(String.self, forKey: .threadStatus)
-		threadActiveFlags = try container.decodeIfPresent([String].self, forKey: .threadActiveFlags) ?? []
-		idleForSeconds = try container.decodeIfPresent(Int.self, forKey: .idleForSeconds)
-		protocolIdleForSeconds = try container.decodeIfPresent(Int.self, forKey: .protocolIdleForSeconds)
-		updatedAt = try container.decodeIfPresent(String.self, forKey: .updatedAt)
-		lastProgressAt = try container.decodeIfPresent(String.self, forKey: .lastProgressAt)
-		nextRetryAt = try container.decodeIfPresent(String.self, forKey: .nextRetryAt)
-		lastEventType = try container.decodeIfPresent(String.self, forKey: .lastEventType)
-		eventCount = try container.decodeIfPresent(Int.self, forKey: .eventCount)
-		executionLiveness = try container.decodeIfPresent(String.self, forKey: .executionLiveness)
-		hasFreshExecutionSnapshot = try container.decodeIfPresent(Bool.self, forKey: .hasFreshExecutionSnapshot)
-		countsAsRunningSnapshot = try container.decodeIfPresent(Bool.self, forKey: .countsAsRunningSnapshot)
-		needsAttentionSnapshot = try container.decodeIfPresent(Bool.self, forKey: .needsAttentionSnapshot)
-		processAlive = try container.decodeIfPresent(Bool.self, forKey: .processAlive)
-		processLivenessReason = try container.decodeIfPresent(String.self, forKey: .processLivenessReason)
-		runLease = try container.decodeIfPresent(Bool.self, forKey: .runLease)
-		branchName = try container.decodeIfPresent(String.self, forKey: .branchName)
-		worktreePath = try container.decodeIfPresent(String.self, forKey: .worktreePath)
-		suspectedStall = try container.decodeIfPresent(Bool.self, forKey: .suspectedStall) ?? false
-		childAgentActivity = try container.decodeIfPresent(
-			OperatorChildAgentActivity.self,
-			forKey: .childAgentActivity
-		)
-		lifecycleMetrics = try container.decodeIfPresent(OperatorLifecycleMetrics.self, forKey: .lifecycleMetrics)
-		continuationRecovery = try container.decodeIfPresent(
-			OperatorContinuationRecoveryStatus.self,
-			forKey: .continuationRecovery
-		)
-		account = try container.decodeIfPresent(OperatorRunAccountSummary.self, forKey: .account)
-			?? container.decodeIfPresent(OperatorRunAccountSummary.self, forKey: .codexAccount)
-		accounts = try container.decodeIfPresent([OperatorRunAccountSummary].self, forKey: .accounts)
-			?? container.decodeIfPresent([OperatorRunAccountSummary].self, forKey: .codexAccounts)
-			?? []
-	}
-
-	private init(
-		projectID: String?,
-		projectDisplayName: String?,
-		runID: String,
-		issueID: String?,
-		issueIdentifier: String?,
-		title: String?,
-		status: String?,
-		attemptStatus: String?,
-		attemptNumber: Int?,
-		phase: String?,
-		runPhase: String?,
-		waitReason: String?,
-		currentOperation: String?,
-		activeGoalPhase: String?,
-		publicProgressPhase: String?,
-		threadStatus: String?,
-		threadActiveFlags: [String],
-		idleForSeconds: Int?,
-		protocolIdleForSeconds: Int?,
-		updatedAt: String?,
-		lastProgressAt: String?,
-		nextRetryAt: String?,
-		lastEventType: String?,
-		eventCount: Int?,
-		executionLiveness: String?,
-		hasFreshExecutionSnapshot: Bool?,
-		countsAsRunningSnapshot: Bool?,
-		needsAttentionSnapshot: Bool?,
-		processAlive: Bool?,
-		processLivenessReason: String?,
-		runLease: Bool?,
-		branchName: String?,
-		worktreePath: String?,
-		suspectedStall: Bool,
-		childAgentActivity: OperatorChildAgentActivity?,
-		lifecycleMetrics: OperatorLifecycleMetrics?,
-		continuationRecovery: OperatorContinuationRecoveryStatus?,
-		account: OperatorRunAccountSummary?,
-		accounts: [OperatorRunAccountSummary]
-	) {
-		self.projectID = projectID
-		self.projectDisplayName = projectDisplayName
-		self.runID = runID
-		self.issueID = issueID
-		self.issueIdentifier = issueIdentifier
-		self.title = title
-		self.status = status
-		self.attemptStatus = attemptStatus
-		self.attemptNumber = attemptNumber
-		self.phase = phase
-		self.runPhase = runPhase
-		self.waitReason = waitReason
-		self.currentOperation = currentOperation
-		self.activeGoalPhase = activeGoalPhase
-		self.publicProgressPhase = publicProgressPhase
-		self.threadStatus = threadStatus
-		self.threadActiveFlags = threadActiveFlags
-		self.idleForSeconds = idleForSeconds
-		self.protocolIdleForSeconds = protocolIdleForSeconds
-		self.updatedAt = updatedAt
-		self.lastProgressAt = lastProgressAt
-		self.nextRetryAt = nextRetryAt
-		self.lastEventType = lastEventType
-		self.eventCount = eventCount
-		self.executionLiveness = executionLiveness
-		self.hasFreshExecutionSnapshot = hasFreshExecutionSnapshot
-		self.countsAsRunningSnapshot = countsAsRunningSnapshot
-		self.needsAttentionSnapshot = needsAttentionSnapshot
-		self.processAlive = processAlive
-		self.processLivenessReason = processLivenessReason
-		self.runLease = runLease
-		self.branchName = branchName
-		self.worktreePath = worktreePath
-		self.suspectedStall = suspectedStall
-		self.childAgentActivity = childAgentActivity
-		self.lifecycleMetrics = lifecycleMetrics
-		self.continuationRecovery = continuationRecovery
-		self.account = account
-		self.accounts = accounts
-	}
-}
-
-struct OperatorContinuationRecoveryStatus: Decodable, Sendable {
-	let state: String?
-	let sourcePhase: String?
-	let nextPhase: String?
-	let sourceErrorClass: String?
-	let sourceErrorMessage: String?
-	let recordedAt: String?
-	let runID: String?
-	let attemptNumber: Int?
-	let recoveryCount: Int?
-	let automaticContinuationLimit: Int?
-	let budgetExceeded: Bool?
-	let nextAction: String?
-
-	enum CodingKeys: String, CodingKey {
-		case state
-		case sourcePhase = "source_phase"
-		case nextPhase = "next_phase"
-		case sourceErrorClass = "source_error_class"
-		case sourceErrorMessage = "source_error_message"
-		case recordedAt = "recorded_at"
-		case runID = "run_id"
-		case attemptNumber = "attempt_number"
-		case recoveryCount = "recovery_count"
-		case automaticContinuationLimit = "automatic_continuation_limit"
-		case budgetExceeded = "budget_exceeded"
-		case nextAction = "next_action"
-	}
-}
-
-private extension Optional where Wrapped == Int {
-	func isSomeAndLessThan(_ threshold: Int) -> Bool {
-		guard let value = self else {
-			return false
-		}
-
-		return value < threshold
-	}
-}
-
-struct OperatorDashboardSocketEvent: Decodable, Sendable {
-	let type: String
-	let payload: OperatorDashboardSocketPayload?
-}
-
-struct OperatorDashboardSocketPayload: Decodable, Sendable {
-	let emittedAtUnixEpoch: Int64?
-	let snapshotPublishedAtUnixEpoch: Int64?
-	let snapshot: OperatorSnapshotResponse?
-	let presentation: OperatorSnapshotPresentation?
-
-	var emittedAt: Date? {
-		date(fromUnixEpoch: emittedAtUnixEpoch)
-	}
-
-	var snapshotPublishedAt: Date? {
-		date(fromUnixEpoch: snapshotPublishedAtUnixEpoch)
-	}
-
-	enum CodingKeys: String, CodingKey {
-		case emittedAtUnixEpoch
-		case snapshotPublishedAtUnixEpoch
-		case snapshot
-		case presentation
-	}
-}
-
-struct OperatorChildAgentActivity: Decodable, Sendable {
-	let currentBucket: String?
-	let currentDetail: String?
-	let currentElapsedSeconds: Int?
-	let currentStartedUnixEpoch: Int64?
-	let eventCount: Int
-	let inputTokensCumulative: Int
-	let inputTokensCurrent: Int?
-	let inputTokensMax: Int?
-	let largestToolOutputBytes: Int?
-	let largestToolOutputTool: String?
-	let outputTokensCumulative: Int
-	let toolCallCount: Int
-	let wallSeconds: Int
-	let buckets: [OperatorChildAgentBucket]
-
-	enum CodingKeys: String, CodingKey {
-		case currentBucket = "current_bucket"
-		case currentDetail = "current_detail"
-		case currentElapsedSeconds = "current_elapsed_seconds"
-		case currentStartedUnixEpoch = "current_started_unix_epoch"
-		case eventCount = "event_count"
-		case inputTokensCumulative = "input_tokens_cumulative"
-		case inputTokensCurrent = "input_tokens_current"
-		case inputTokensMax = "input_tokens_max"
-		case largestToolOutputBytes = "largest_tool_output_bytes"
-		case largestToolOutputTool = "largest_tool_output_tool"
-		case outputTokensCumulative = "output_tokens_cumulative"
-		case toolCallCount = "tool_call_count"
-		case wallSeconds = "wall_seconds"
-		case buckets
-	}
-
-	init(from decoder: Decoder) throws {
-		let container = try decoder.container(keyedBy: CodingKeys.self)
-
-		currentBucket = try container.decodeIfPresent(String.self, forKey: .currentBucket)
-		currentDetail = try container.decodeIfPresent(String.self, forKey: .currentDetail)
-		currentElapsedSeconds = try container.decodeIfPresent(Int.self, forKey: .currentElapsedSeconds)
-		currentStartedUnixEpoch = try container.decodeIfPresent(Int64.self, forKey: .currentStartedUnixEpoch)
-		eventCount = try container.decodeIfPresent(Int.self, forKey: .eventCount) ?? 0
-		inputTokensCumulative = try container.decodeIfPresent(Int.self, forKey: .inputTokensCumulative) ?? 0
-		inputTokensCurrent = try container.decodeIfPresent(Int.self, forKey: .inputTokensCurrent)
-		inputTokensMax = try container.decodeIfPresent(Int.self, forKey: .inputTokensMax)
-		largestToolOutputBytes = try container.decodeIfPresent(Int.self, forKey: .largestToolOutputBytes)
-		largestToolOutputTool = try container.decodeIfPresent(String.self, forKey: .largestToolOutputTool)
-		outputTokensCumulative = try container.decodeIfPresent(Int.self, forKey: .outputTokensCumulative) ?? 0
-		toolCallCount = try container.decodeIfPresent(Int.self, forKey: .toolCallCount) ?? 0
-		wallSeconds = try container.decodeIfPresent(Int.self, forKey: .wallSeconds) ?? 0
-		buckets = try container.decodeIfPresent([OperatorChildAgentBucket].self, forKey: .buckets) ?? []
-	}
-
-	func currentElapsedSeconds(at now: Date) -> Int? {
-		var candidates = [Int]()
-		if let currentElapsedSeconds {
-			candidates.append(currentElapsedSeconds)
-		}
-		if let currentStartedUnixEpoch {
-			let liveElapsed = Int(now.timeIntervalSince1970.rounded(.down)) - Int(currentStartedUnixEpoch)
-
-			candidates.append(max(0, liveElapsed))
-		}
-
-		return candidates.max()
-	}
-
-	func wallSeconds(at now: Date) -> Int {
-		wallSeconds + currentElapsedDelta(at: now)
-	}
-
-	func wallSeconds(
-		for bucket: OperatorChildAgentBucket,
-		at now: Date
-	) -> Int {
-		guard let currentBucket, bucket.name.caseInsensitiveCompare(currentBucket) == .orderedSame else {
-			return bucket.wallSeconds
-		}
-
-		return bucket.wallSeconds + currentElapsedDelta(at: now)
-	}
-
-	private func currentElapsedDelta(at now: Date) -> Int {
-		guard let baselineElapsed = currentElapsedSeconds, let liveElapsed = currentElapsedSeconds(at: now) else {
-			return 0
-		}
-
-		return max(0, liveElapsed - baselineElapsed)
-	}
-}
-
-struct OperatorChildAgentBucket: Decodable, Identifiable, Sendable {
-	let name: String
-	let eventCount: Int
-	let inputTokens: Int
-	let outputBytes: Int
-	let outputTokens: Int
-	let toolCallCount: Int
-	let wallSeconds: Int
-
-	var id: String {
-		name
-	}
-
-	enum CodingKeys: String, CodingKey {
-		case name
-		case eventCount = "event_count"
-		case inputTokens = "input_tokens"
-		case outputBytes = "output_bytes"
-		case outputTokens = "output_tokens"
-		case toolCallCount = "tool_call_count"
-		case wallSeconds = "wall_seconds"
-	}
-
-	init(from decoder: Decoder) throws {
-		let container = try decoder.container(keyedBy: CodingKeys.self)
-
-		name = try container.decodeIfPresent(String.self, forKey: .name) ?? "Activity"
-		eventCount = try container.decodeIfPresent(Int.self, forKey: .eventCount) ?? 0
-		inputTokens = try container.decodeIfPresent(Int.self, forKey: .inputTokens) ?? 0
-		outputBytes = try container.decodeIfPresent(Int.self, forKey: .outputBytes) ?? 0
-		outputTokens = try container.decodeIfPresent(Int.self, forKey: .outputTokens) ?? 0
-		toolCallCount = try container.decodeIfPresent(Int.self, forKey: .toolCallCount) ?? 0
-		wallSeconds = try container.decodeIfPresent(Int.self, forKey: .wallSeconds) ?? 0
-	}
-}
-
-struct OperatorRunAccountSummary: Decodable, Sendable {
-	let accountFingerprint: String
-	let email: String?
-	let status: String?
-
-	var isSelected: Bool {
-		status?.caseInsensitiveCompare("selected") == .orderedSame
-	}
-
-	func matches(_ account: CodexAccount) -> Bool {
-		if accountFingerprint.isEmpty == false, accountFingerprint == account.accountFingerprint {
-			return true
-		}
-		if let email, let accountEmail = account.email {
-			return email.caseInsensitiveCompare(accountEmail) == .orderedSame
-		}
-
-		return false
-	}
-
-	enum CodingKeys: String, CodingKey {
-		case accountFingerprint = "account_fingerprint"
-		case email
-		case accountEmail = "account_email"
-		case status
-	}
-
-	init(from decoder: Decoder) throws {
-		let container = try decoder.container(keyedBy: CodingKeys.self)
-
-		accountFingerprint = try container.decodeIfPresent(String.self, forKey: .accountFingerprint) ?? ""
-		email = try container.decodeIfPresent(String.self, forKey: .email)
-			?? container.decodeIfPresent(String.self, forKey: .accountEmail)
-		status = try container.decodeIfPresent(String.self, forKey: .status)
-	}
-}
-
-struct OperatorSnapshotPresentation: Decodable, Sendable {
-	let schema: String?
-	let currentLaneCards: [OperatorCurrentLaneCard]
-
-	enum CodingKeys: String, CodingKey {
-		case schema
-		case currentLaneCards = "current_lane_cards"
-	}
-
-	init(from decoder: Decoder) throws {
-		let container = try decoder.container(keyedBy: CodingKeys.self)
-
-		schema = try container.decodeIfPresent(String.self, forKey: .schema)
-		currentLaneCards = try container.decodeIfPresent(
-			[OperatorCurrentLaneCard].self,
-			forKey: .currentLaneCards
-		) ?? []
-	}
-}
-
-struct OperatorCurrentLaneCard: Decodable, Identifiable, Sendable {
-	let id: String
-	let runID: String
-	let issueID: String?
-	let issueIdentifier: String?
-	let projectID: String?
-	let title: String
-	let detail: String
-	let tone: String
-	let countsAsRunning: Bool
-	let needsAttention: Bool
-	let isWaiting: Bool
-	let assignedAccountFingerprints: [String]
-	let assignedAccountEmails: [String]
-	let run: OperatorRunStatus
-
-	func isAssigned(to account: CodexAccount) -> Bool {
-		if assignedAccountFingerprints.contains(where: { $0 == account.accountFingerprint }) {
-			return true
-		}
-		guard let accountEmail = account.email else {
-			return false
-		}
-
-		return assignedAccountEmails.contains {
-			$0.caseInsensitiveCompare(accountEmail) == .orderedSame
-		}
-	}
-
-	enum CodingKeys: String, CodingKey {
-		case id
-		case runID = "run_id"
-		case issueID = "issue_id"
-		case issueIdentifier = "issue_identifier"
-		case projectID = "project_id"
-		case title
-		case detail
-		case tone
-		case countsAsRunning = "counts_as_running"
-		case needsAttention = "needs_attention"
-		case isWaiting = "is_waiting"
-		case assignedAccountFingerprints = "assigned_account_fingerprints"
-		case assignedAccountEmails = "assigned_account_emails"
-		case run
-	}
-}
-
-private func trimmed(_ value: String?) -> String? {
-	value?.trimmingCharacters(in: .whitespacesAndNewlines)
-}
-
-private func rawDisplayToken(_ value: String) -> String {
-	value.trimmingCharacters(in: .whitespacesAndNewlines)
-}
-
-private func formatOperatorActivityDuration(_ seconds: Int?) -> String? {
-	guard let seconds else {
-		return nil
-	}
-
-	let value = max(0, seconds)
-	if value < 60 {
-		return "\(value)s"
-	}
-
-	let hours = value / 3_600
-	let minutes = (value % 3_600) / 60
-	let remainderSeconds = value % 60
-	if hours > 0 {
-		return minutes > 0 ? "\(hours)h \(minutes)m" : "\(hours)h"
-	}
-	if minutes > 0 {
-		return remainderSeconds > 0 ? "\(minutes)m \(remainderSeconds)s" : "\(minutes)m"
-	}
-
-	return "\(remainderSeconds)s"
-}
-
-private func formatOperatorCompactCount(_ value: Int) -> String {
-	let absoluteValue = abs(Double(value))
-	let sign = value < 0 ? "-" : ""
-
-	if absoluteValue >= 1_000_000_000 {
-		return "\(sign)\(formatOperatorCompactDecimal(absoluteValue / 1_000_000_000))B"
-	}
-	if absoluteValue >= 1_000_000 {
-		return "\(sign)\(formatOperatorCompactDecimal(absoluteValue / 1_000_000))M"
-	}
-	if absoluteValue >= 1_000 {
-		return "\(sign)\(formatOperatorCompactDecimal(absoluteValue / 1_000))k"
-	}
-
-	return "\(value)"
-}
-
-private func formatOperatorCompactDecimal(_ value: Double) -> String {
-	if value >= 100 {
-		return String(format: "%.0f", value)
-	}
-	if value >= 10 {
-		return String(format: "%.1f", value)
-	}
-
-	return String(format: "%.2f", value)
-}
-
-private func formatOperatorCompactBytes(_ value: Int) -> String {
-	let units = ["B", "KiB", "MiB", "GiB"]
-	var amount = Double(max(0, value))
-	var unitIndex = 0
-	while amount >= 1024, unitIndex < units.count - 1 {
-		amount /= 1024
-		unitIndex += 1
-	}
-
-	if unitIndex == 0 {
-		return "\(Int(amount))\(units[unitIndex])"
-	}
-	if amount >= 100 {
-		return "\(Int(amount.rounded()))\(units[unitIndex])"
-	}
-
-	return String(format: "%.1f%@", amount, units[unitIndex])
-}
-
-private func date(fromUnixEpoch value: Int64?) -> Date? {
-	guard let value else {
-		return nil
-	}
-
-	return Date(timeIntervalSince1970: TimeInterval(value))
 }
